inranet_search/requirements.txt
anyio==3.6.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
fastapi==0.85.1
h11==0.14.0
httptools==0.5.0
idna==3.4
packaging==21.3
Pillow==9.2.0
pydantic==1.10.2
pymongo==4.3.2
pyparsing==3.0.9
pytesseract==0.3.10
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tika==1.24
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.0
websockets==10.4
